requirements.txt
pyserial==3.5
pyyaml==5.4.1
numpy==1.24.3
lewansoul_lx16a==0.1